Add encoding_for_model method
